+++ node/pkg/query/query.go
@@ -129,6 +129,7 @@ var perChainConfig = map[vaa.ChainID]PerChainConfig{
 	vaa.ChainIDBlast:           {NumWorkers: 1, TimestampCacheSupported: true},
 	vaa.ChainIDXLayer:          {NumWorkers: 1, TimestampCacheSupported: true},
 	vaa.ChainIDLinea:           {NumWorkers: 1, TimestampCacheSupported: true},
+	vaa.ChainIDBerachain:       {NumWorkers: 1, TimestampCacheSupported: true},
 	vaa.ChainIDSepolia:         {NumWorkers: 1, TimestampCacheSupported: true},
 	vaa.ChainIDHolesky:         {NumWorkers: 1, TimestampCacheSupported: true},
 	vaa.ChainIDArbitrumSepolia: {NumWorkers: 1, TimestampCacheSupported: true},