+++ sdk/js/CHANGELOG.md
@@ -6,6 +6,7 @@ Support for move of relayer ethereum contracts
 Blast Sepolia support
 XLayer Sepolia support
 Linea Sepolia support
+Berachain Testnet support
 
 ## 0.10.13